LayoutTests/platform/mac-wk1/imported/w3c/web-platform-tests/html/browsers/windows/clear-window-name.https-expected.txt

Blocked access to external URL https://www.localhost:9443/html/browsers/windows/resources/window-name.sub.html?set=198786cf-d727-411f-bc70-1e58e1985bd8%7CsetDomain=localhost:9443%7Csub%7Creport=198786cf-d727-411f-bc70-1e58e1985bd8%7Cclose

Harness Error (TIMEOUT), message = null

PASS Window.name is not reset when there is an opener around
PASS Window.name is not reset when there is an opener around (cross-origin)
PASS Window.name is not reset at the first navigation away from initial about:blank with noopener
PASS Window.name is reset at the first cross-origin navigation with noopener
PASS Window.name is not reset at the first navigation away from initial about:blank with window.opener set to null
PASS Window.name is not reset at the same-origin navigation with window.opener set to null
PASS Window.name is reset at the first first cross-origin navigation with window.opener set to null
PASS Window.name is set by the window
PASS Window.name is reset at the first cross-origin navigation
PASS window.name is not reset after navigating to an about:blank page from a non-about:blank page
TIMEOUT Window.name is not reset if the document.domain is set to the parent domain Test timed out
